+++ src/@cdk/components/tarefa/cdk-tarefa-grid/cdk-tarefa-grid.component.ts
@@ -223,6 +223,9 @@ export class CdkTarefaGridComponent implements AfterViewInit, OnInit, OnChanges 
     @Output()
     selectedIds: number[] = [];
 
+    @Output()
+    listCompartilhamentos = new EventEmitter<number>();
+
     dataSource: TarefaDataSource;
 
     showFilter = false;
@@ -343,6 +346,10 @@ export class CdkTarefaGridComponent implements AfterViewInit, OnInit, OnChanges 
 
     listAtividade(tarefaId): void {
         this.listAtividades.emit(tarefaId);
+    }
+
+    listCompartilhamento(tarefaId): void {
+        this.listCompartilhamentos.emit(tarefaId);
     }
 
     deleteBloco(ids): void {